% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_text.R
\name{extract_text}
\alias{extract_text}
\title{extract_text}
\usage{
extract_text(img_file)
}
\arguments{
\item{img_file}{Screenshot of the schedule for the week}
}
\value{
Text extracted from the screenshot
}
\description{
Extract text from a screenshot image
}
\examples{
img_file <- system.file("testdata/IMG_0579.PNG", package = "ScreenShot2CalendarEndpoint")
extract_text(img_file)
}
