% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_shifts.R
\name{get_shifts}
\alias{get_shifts}
\title{get_shifts}
\usage{
get_shifts(lines)
}
\arguments{
\item{lines}{Lines extracted from \code{\link{extract_text}}}
}
\value{
Data frame where each row is a shift
}
\description{
Description
}
\details{
Get each shift in a tidy format
}
\examples{
\dontrun{text \%>\% 
    str_split_1("\n") \%>\% 
    get_shifts()}
}
