% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_week.R
\name{get_week}
\alias{get_week}
\title{get_week}
\usage{
get_week(lines)
}
\arguments{
\item{lines}{Lines extracted from \code{\link{extract_text}}}
}
\value{
Start and end dates of the week
}
\description{
Description
}
\details{
Get the week start and end dates from the converted text
}
\examples{
\dontrun{
img_file <- system.file("testdata/IMG_0579.PNG", package = "ScreenShot2CalendarEndpoint")
text <- extract_text(img_file)
text \%>\% 
  str_split_1("\n") \%>\% 
  get_week()
}

}
