% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScreenShot2CalendarEndpoint-package.R
\docType{package}
\name{ScreenShot2CalendarEndpoint-package}
\alias{ScreenShot2CalendarEndpoint}
\alias{ScreenShot2CalendarEndpoint-package}
\title{ScreenShot2CalendarEndpoint: ScreenShot2CalendarEndpoint}
\description{
Use a screenshot of the schedule to generate an ical event file using an R API endpoint.
}
\author{
\strong{Maintainer}: Tinashe Tapera \email{tinashemtapera@gmail.com}

}
\keyword{internal}
